fix(switch): fix accessibility issues
putting the role in the right place
association aria and data attributes with the container
using native attributes (checked,required,disabled) on hidden input

here is where it gets fun:
identified issue where checked is not applied correctly and found this gem:
https://github.com/angular/components/issues/26101
current approach uses preventDefault in toggle and breaks.
this current version works but not if we click on the switch (background) and not the thumb
what's even more fun: e2e tests pass because cypress clicks right in the middle where our thumb is.

open to suggestions and help!

fixes #648